requirements.txt
streamlit
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1
faiss-cpu
python-dotenv
docx2txt
pypdf
langchain-text-splitters
sentence-transformers
langchain-pinecone
pinecone-client
